backend/pyproject.toml
[project]
name = "peluprice-backend"
version = "0.1.0"
description = "PeluPrice IoT Device Management Backend"
authors = [
    { name = "PeluPrice Team", email = "team@peluprice.com" }
]
requires-python = ">=3.9"

dependencies = [
    "fastapi>=0.103.2",
    "uvicorn[standard]>=0.23.2",
    "sqlalchemy>=2.0.22",
    "psycopg2-binary>=2.9.7",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "websockets>=11.0.3",
    "paho-mqtt>=1.6.1",
    "redis>=4.6.0",
    "python-dotenv>=1.0.0",
    "alembic>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"